+++ tests/Feature/AngleTest.php
@@ -409,9 +409,31 @@ class AngleTest extends TestCase
     #[TestDox("can toggle its direction.")]
     public function test_can_toggle_rotation_direction(): void
     {
+        /**
+         * With SexadecimalDegrees
+         */
         // Arrange
+        $failure_message_1 = "The angle should be counterclockwise but found the opposite.";
+        $failure_message_2 = "The angle should be clockwise but found the opposite.";
         $alfa = $this->positiveRandomAngle();
         $beta = $this->negativeRandomAngle();
+
+        // Act & Assert
+        $this->assertDirection(
+            $alfa->direction->opposite(), 
+            $alfa->toggleDirection()->direction, 
+            $failure_message_2
+        );
+        $this->assertDirection(
+            $beta->direction->opposite(), 
+            $beta->toggleDirection()->direction, 
+            $failure_message_1
+        );
+
+        /**
+         * With SexadecimalDegrees
+         */
+        // Arrange
         $gamma = Angle::createFromValues(
             $this->randomDegrees()->value(), 
             direction: Direction::COUNTER_CLOCKWISE
@@ -421,31 +443,15 @@ class AngleTest extends TestCase
             direction: Direction::CLOCKWISE
         );
 
-        // Act
-        //  With SexadecimalDegrees
-        $alfa_opposite = $alfa->toggleDirection();      // From positive to negative
-        $beta_opposite = $beta->toggleDirection();      // From negative to positive
-        //  Without SexadecimalDegrees
-        $gamma_opposite = $gamma->toggleDirection();    // From positive to negative 
-        $delta_opposite = $delta->toggleDirection();    // From negative to positive
-
-        // Assert
-        $failure_message_1 = "The angle should be counterclockwise but found the opposite.";
-        $failure_message_2 = "The angle should be clockwise but found the opposite.";
-        //  With SexadecimalDegrees
-        $this->assertDirection(Direction::CLOCKWISE, $alfa_opposite->direction, $failure_message_2);
-        $this->assertEquals(
-            $alfa->toSexadecimalDegrees()->value->mul(-1)->toFloat(), 
-            $alfa_opposite->toSexadecimalDegrees()->value()
+        // Act & Assert
+        $this->assertSexadecimalDegrees(
+            $gamma->toSexadecimalDegrees()->toggleDirection(),
+            $gamma->toggleDirection()->toSexadecimalDegrees()
         );
-        $this->assertDirection(Direction::COUNTER_CLOCKWISE, $beta_opposite->direction, $failure_message_1);
-        $this->assertEquals(
-            $beta->toSexadecimalDegrees()->value->mul(-1)->toFloat(),
-            $beta_opposite->toSexadecimalDegrees()->value()
+        $this->assertSexadecimalDegrees(
+            $delta->toSexadecimalDegrees()->toggleDirection(),
+            $delta->toggleDirection()->toSexadecimalDegrees()
         );
-        // Without SexadecimalDegrees
-        $this->assertDirection(Direction::CLOCKWISE, $gamma_opposite->direction, $failure_message_2);
-        $this->assertDirection(Direction::COUNTER_CLOCKWISE, $delta_opposite->direction, $failure_message_1);
     }
 
     #[TestDox("can be equal compared against an int, float, string or Angle.")]